# Conflicts:
#	crates/sui-framework/docs/bfc-system/bfc_system.md
#	crates/sui-framework/docs/bridge/bridge.md
#	crates/sui-framework/docs/bridge/chain_ids.md
#	crates/sui-framework/docs/bridge/limiter.md
#	crates/sui-framework/docs/bridge/message.md
#	crates/sui-framework/docs/bridge/message_types.md
#	crates/sui-framework/docs/bridge/treasury.md
#	crates/sui-graphql-e2e-tests/tests/stable/available_range/available_range.exp
#	crates/sui-graphql-e2e-tests/tests/stable/call/checkpoint_connection_pagination.exp
#	crates/sui-graphql-e2e-tests/tests/stable/call/dynamic_fields.exp
#	crates/sui-graphql-e2e-tests/tests/stable/call/simple.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/balances.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/checkpoints/transaction_blocks.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/coins.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/deleted_df.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/deleted_dof.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/dof_add_reclaim_transfer.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/dof_add_reclaim_transfer_reclaim_add.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/dynamic_fields.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/immutable_dof.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/mutated_df.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/mutated_dof.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/dynamic_fields/nested_dof.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/epochs/checkpoints.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/epochs/transaction_blocks.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/object_at_version.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/objects_pagination.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/objects_pagination_single.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/performance/many_objects.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/staked_sui.exp
#	crates/sui-graphql-e2e-tests/tests/stable/consistency/tx_address_objects.exp
#	crates/sui-graphql-e2e-tests/tests/stable/epoch/chain_identifier.exp
#	crates/sui-graphql-e2e-tests/tests/stable/epoch/epoch.exp
#	crates/sui-graphql-e2e-tests/tests/stable/epoch/epoch_start_to_end.exp
#	crates/sui-graphql-e2e-tests/tests/stable/errors/clever_errors.exp
#	crates/sui-graphql-e2e-tests/tests/stable/errors/clever_errors_in_macros.exp
#	crates/sui-graphql-e2e-tests/tests/stable/event_connection/event_connection.exp
#	crates/sui-graphql-e2e-tests/tests/stable/event_connection/nested_emit_event.exp
#	crates/sui-graphql-e2e-tests/tests/stable/event_connection/pagination.exp
#	crates/sui-graphql-e2e-tests/tests/stable/event_connection/tx_digest.exp
#	crates/sui-graphql-e2e-tests/tests/stable/event_connection/type_filter.exp
#	crates/sui-graphql-e2e-tests/tests/stable/event_connection/type_param_filter.exp
#	crates/sui-graphql-e2e-tests/tests/stable/events/sending_module.exp
#	crates/sui-graphql-e2e-tests/tests/stable/limits/directives.exp
#	crates/sui-graphql-e2e-tests/tests/stable/limits/output_node_estimation.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/coin.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/data.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/display.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/enum_data.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/full_objects_history.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/pagination.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/public_transfer.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/received.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/staked_sui.exp
#	crates/sui-graphql-e2e-tests/tests/stable/objects/wrap_unwrap.exp
#	crates/sui-graphql-e2e-tests/tests/stable/owner/downcasts.exp
#	crates/sui-graphql-e2e-tests/tests/stable/owner/root_version.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/bcs.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/datatypes.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/enums.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/friends.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/functions.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/modules.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/structs.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/types.exp
#	crates/sui-graphql-e2e-tests/tests/stable/packages/versioning.exp
#	crates/sui-graphql-e2e-tests/tests/stable/prune/prune.exp
#	crates/sui-graphql-e2e-tests/tests/stable/prune/transactions.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transaction_block_effects/balance_changes.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transaction_block_effects/dependencies.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transaction_block_effects/events.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/at_checkpoint.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/errors.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/filters/kind.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/programmable.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/alternating.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/both_cursors.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/equal/first.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/equal/last.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/ge_page/first.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/ge_page/last.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/le_page/first.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/le_page/last.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/scan_limit/require.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/shared.exp
#	crates/sui-graphql-e2e-tests/tests/stable/transactions/system.exp
#	crates/sui-swarm-config/tests/snapshots/snapshot_tests__populated_genesis_snapshot_matches-2.snap

+++ bridge/evm/contracts/BridgeConfig.sol
@@ -31,7 +31,7 @@ contract BridgeConfig is IBridgeConfig, CommitteeUpgradeable {
         uint8 _chainID,
         address[] memory _supportedTokens,
         uint64[] memory _tokenPrices,
-        uint8[] memory _tokenIds,
+        uint64[] memory _tokenIds,
         uint8[] memory _suiDecimals,
         uint8[] memory _supportedChains
     ) external initializer {
